Add -f option, to fork after a successful locking.
This permits a chain of commands of the form 'xtrlock -f && suspend',
where 'suspend' is whatever CLI command suspends your particular
laptop environment, so that when you unsuspend you then have to unlock
the xtrlock.

Without this option, you have to run xtrlock in the background, and
then either include a bodgy sleep to try to avoid the suspend command
winning the race, or else find a way to test by some X-related means
that the locking has happened.

Signed-off-by: Simon Tatham <anakin@pobox.com>